♻️ Move icons to single file
